Game progress is saved using levelprogress which uses progresscontroller.
Currently the game progress gets saved (was answer right or wrong, score, current clicks in the level aka where the user currently is at the level). They are saved to shared preferences, and after each rounds gets 10 clicks the user gets the next level and saving to that levels "memory" starts instead of the previous ones. The progress bar in the game view shows correctly the current and past progress in each level even if the user reopens the app, the current progress is reloaded from memory.