Level activity and play button should now direct to game level correctly
When pressing the play button in main menu or a level in the level activity, the player should now be directed to the correct level as wanted, and continued to redirect to the correct level when continuing to play the game. The unlocked levels should also be greyed out and disabled in the levels activity until the level has been reached. If a level has been played to 10 clicks, it will be reset upon playing again.