Reset game progress at appropriate times.
Reset the current game progress when clicking play through the menu and that current level has not been played last, and also reset the game if the new level is different from the last one, so if the player has played a little of level 3 before and then plays level 2 and goes through to level 3, the level is reset. Only save the progress of the max level played (through play button) and if the clicked level through the play button or if the the player has been playing some level through the levels activity and it's the current level last played.